fix to work with GWE in vite
